% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buffer.R
\name{buffer_ppn}
\alias{buffer_ppn}
\title{buffer_ppn}
\usage{
buffer_ppn(ppn, buffer_size)
}
\arguments{
\item{ppn}{A spatial point dataframe}

\item{buffer_size}{A vector of integers}
}
\value{
Spatial polygons
}
\description{
Generate buffers
}
\author{
Giampaolo Cocca
}

