% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_map.R
\name{draw_map}
\alias{draw_map}
\title{Draw a map of PPNs over Sweden}
\usage{
draw_map(ppn, ppn_obj = system.file("extdata/result.rda", package = "svamp"))
}
\arguments{
\item{ppn_obj}{The list of dataset output from package SVDC}

\item{outbreak}{Scalar or numeric vector with the PPNs number to draw}
}
\value{
A .png map
}
\description{
Draw a map of outbreak points over the Län of Sweden.
}
\author{
Giampaolo Cocca
}

