% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buffer.R
\name{hole}
\alias{hole}
\title{hole}
\usage{
hole(buffer_polygons, buffer_size)
}
\arguments{
\item{buffer_polygons}{A spatial polygons object}

\item{buffer_size}{A vector of integers}
}
\value{
spatial polygon dataframe
}
\description{
Generate a polygon set with holes
}
\author{
Giampaolo Cocca
}

