% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/svamp.R
\name{report}
\alias{report}
\title{Produce an .html report}
\usage{
report(ppn = 222240, firstname = "Giampaolo", lastname = "Cocca",
  template, format = c("knitr"))
}
\arguments{
\item{ppn}{ppn number (single scalar value)}

\item{firstname}{Firstname of the person running the report}

\item{lastname}{Lastname of the person running the report}

\item{ppn_obj}{Path to the list of objects output from svdc package}

\item{X}{X-coordinate of the outbreak provided by the user when ppn coordinates are missing}

\item{Y}{Y-coordinate of the outbreak provided by the user when ppn coordinates are missing}

\item{buf1}{Size in meters of the first buffer drawn around the ppn}

\item{buf2}{Size in meters of the second buffer drawn around the ppn}

\item{buf3}{Size in meters of the third buffer drawn around the ppn}
}
\value{
An html report
}
\description{
Produce an .html report with information for outbreak/suspicion management
}

