\name{svamp-package}
\alias{svamp-package}
\alias{svamp}
\docType{package}
\title{
SVA mapping application (SVAMP)
}
\description{
Takes a dataset cleaned by the svdc package and produces a map report
depending on the PPN given to the report function
}
\details{
\tabular{ll}{
Package: \tab svamp\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-09-07\cr
License: \tab Internal SVA code\cr
}
}
\author{
Giampaolo Cocca
}
\references{
None
}
\keyword{ package }
\seealso{
None
}
\examples{
None
}
