% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{DistMatrix}
\alias{DistMatrix}
\title{DistMatrix}
\usage{
DistMatrix(farms_in_buffers, longlat = FALSE, outbreak)
}
\arguments{
\item{farms_in_buffers}{A spatial point data.frame of farms}

\item{longlat}{coordinate type}

\item{outbreak}{the reference farm(s)}
}
\value{
Matrix
}
\description{
Calculate the distance between farms
}
\author{
Giampaolo Cocca
}

