% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/report.R
\name{draw_map}
\alias{draw_map}
\title{Draw a map}
\usage{
draw_map(ppn, ppn_obj = system.file("extdata/result.rda", package = "svamp"))
}
\arguments{
\item{outbreak}{Scalar or numeric vector with the PPNs number to draw}
}
\value{
A .jpeg? map
}
\description{
Draw a map of outbreak points over the Län of Sweden.
}
\keyword{internal}

