% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{report}
\alias{report}
\title{Produce an .html report}
\usage{
report(ppn, ppn_obj = system.file("extdata/result.rda", package = "svamp"),
  firstname = "", lastname = "", buffer_size = c(3, 10), days = 90,
  template = "report", format = c("knitr"), view = FALSE)
}
\arguments{
\item{ppn}{Ppn numbers (vector of int numbers, comma separeted)}

\item{ppn_obj}{Path to the list of objects output from svdc package}

\item{firstname}{Firstname of the person running the report}

\item{lastname}{Lastname of the person running the report}

\item{buffer_size}{Size in kilometers of the buffers drawn around the ppn (vector of numbers, comma separeted)}

\item{days}{Set the number of days to use in EpiContactTrace (max 180)}

\item{view}{Make TRUE to pop a browser}

\item{X}{X-coordinate of the outbreak provided by the user when ppn coordinates are missing}

\item{Y}{Y-coordinate of the outbreak provided by the user when ppn coordinates are missing}
}
\value{
An html report
}
\description{
Produce an .html report with information for outbreak/suspicion management
}

